requirements.txt
allure-pytest==2.9.42
allure-python-commons==2.9.42
Appium-Python-Client==1.1.0
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
confparser==0.1.7
et-xmlfile==1.1.0
glob2==0.7
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==2.0.1
openpyxl==3.0.7
packaging==20.9
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-rerunfailures==10.0
PyYAML==5.3.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
